Weights are double not float. Using float causes overflow for large graphs

+++ CE_Louvain_V3/community.cpp
@@ -81,7 +81,7 @@ Community::init_partition(char * filename) {
       unsigned int i=0;
       for ( i=0 ; i<neigh_last ; i++) {
 	unsigned int best_comm     = neigh_pos[i];
-	float best_nblinks  = neigh_weight[neigh_pos[i]];
+	double best_nblinks  = neigh_weight[neigh_pos[i]];
 	if (best_comm==comm) {
 	  insert(node, best_comm, best_nblinks);
 	  break;
@@ -139,7 +139,7 @@ Community::neigh_comm(unsigned int node) {
     neigh_weight[neigh_pos[i]]=-1;
   neigh_last=0;
 
-  pair<vector<unsigned int>::iterator, vector<float>::iterator> p = g.neighbors(node);
+  pair<vector<unsigned int>::iterator, vector<double>::iterator> p = g.neighbors(node);
 
   unsigned int deg = g.nb_neighbors(node);
 
@@ -176,7 +176,7 @@ Community::partition2graph() {
 
 
   for (int i=0 ; i<size ; i++) {
-    pair<vector<unsigned int>::iterator, vector<float>::iterator> p = g.neighbors(i);
+    pair<vector<unsigned int>::iterator, vector<double>::iterator> p = g.neighbors(i);
 
     int deg = g.nb_neighbors(i);
     for (int j=0 ; j<deg ; j++) {
@@ -257,7 +257,7 @@ Community::reorder() {
     
     // Iterate over members of communities
     for (int node=0 ; node<comm_size ; node++) {
-      pair<vector<unsigned int>::iterator, vector<float>::iterator> p = g.neighbors(comm_nodes[comm][node]);
+      pair<vector<unsigned int>::iterator, vector<double>::iterator> p = g.neighbors(comm_nodes[comm][node]);
       int deg = g.nb_neighbors(comm_nodes[comm][node]);
       
       // Commulative sum of degrees 
@@ -304,12 +304,12 @@ Community::partition2graph_binary() {
 
   int comm_deg = comm_nodes.size();
   for (int comm=0 ; comm<comm_deg ; comm++) {
-    map<int,float> m;
-    map<int,float>::iterator it;
+    map<int,double> m;
+    map<int,double>::iterator it;
 
     int comm_size = comm_nodes[comm].size();
     for (int node=0 ; node<comm_size ; node++) {
-      pair<vector<unsigned int>::iterator, vector<float>::iterator> p = g.neighbors(comm_nodes[comm][node]);
+      pair<vector<unsigned int>::iterator, vector<double>::iterator> p = g.neighbors(comm_nodes[comm][node]);
       int deg = g.nb_neighbors(comm_nodes[comm][node]);
       for (int i=0 ; i<deg ; i++) {
 	int neigh        = *(p.first+i);
